clean architecture 리팩토링

+++ domain/src/main/java/com/example/domain/model/Location.kt
@@ -1,6 +1,6 @@
 package com.example.domain.model
 
 data class Location(
-    val x: Int,
-    val y: Int
+    val x: Double,
+    val y: Double
 )
